chore: merge sync stats with metrics file
